+++ controllers/routers/posts.js
@@ -36,11 +36,11 @@ exports.editor.edit = {
         return next();
     },
     patch: async (req, res) => {
-        const doc = await postsModel.findByIdAndUpdate(_U_.string.readMongoId(req.url), req.body.post, { new: true });
+        await postsModel.update({ _id: _U_.string.readMongoId(req.url)}, { $set: req.body.post }, { new: true });
         req.flash('info', 'post have been successfully updated!');
-        return res.redirect(`/posts/${doc.canonical}`);
+        return res.redirect(`/posts/${_U_.string.readMongoId(req.url)}`);
     },
-    delete: async (req, res) => { // todo: trash can || double check
+    delete: async (req, res) => {
         await postsModel.remove({ _id: _U_.string.readMongoId(req.url) });
         req.flash('info', 'post have been successfully deleted!');
         return res.redirect('/posts/');
@@ -48,16 +48,18 @@ exports.editor.edit = {
 };
 
 exports.posts.show = {
-    alias: async (req, res) => {
-        if (req.session.user) req.session.chest = { post : await postsModel.findOne(req.params) };
-        else req.session.chest = { post : await postsModel.findOne({ ...req.params, status: 'published' })};
+    alias: async (req, res) => {                                                                                        // tofix: revise alias behavior (merged into `postHandler`)
+        const post = await postsModel.findOne({ ...req.params, 'time._recycled': { $eq: null }});
+        if (!post) throw new _U_.error.HttpError(404);
+        else req.session.chest = { post };
         return res.redirect(`/posts/${req.session.chest.post.canonical}`);
     },
-    get: async (req, res, next) => {    // tofix: post not found page
+    get: async (req, res, next) => {                                                                                    // tofix: post not found page
         if (!req.session.chest) {
-            const query = { canonical: req.params[0] };
-            if (req.session.user) req.session.chest = { post : await postsModel.findOne(query)};
-            else req.session.chest = { post : await postsModel.findOne({ ...query, status: 'published' })};
+            console.log(123);
+            req.session.chest = {
+                post : await postsModel.findOne({ canonical: req.params[0], 'time._recycled': { $eq: null }})
+            };
         } return next();
     },
 };