+++ models/index.js
@@ -1,6 +1,6 @@
 module.exports = {
-    configsModel: require('./configs'),
-    usersModel  : require('./users'),
-    postsModel  : require('./posts'),
-    mediaModel  : require('./media'),
+    configsModel    : require('./configs'),
+    usersModel      : require('./users'),
+    postsModel      : require('./posts'),
+    mediaModel      : require('./media'),
 };